requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-import-export==3.0.1
django-simple-history==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environ==1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
